requirements.txt
nltk==3.9.1
pymongo~=4.13.2
python-dotenv==1.1.0
structlog==25.2.0